+++ message-consumer-app/src/main/java/se/npet/microservices/messageconsumer/service/TemperatureDeviationCalculatorImpl.java
@@ -1,5 +1,6 @@
 package se.npet.microservices.messageconsumer.service;
 
+import java.math.BigDecimal;
 import org.springframework.stereotype.Component;
 import se.npet.microservices.messageconsumer.service.model.Deviation;
 
@@ -7,7 +8,7 @@ import se.npet.microservices.messageconsumer.service.model.Deviation;
 public class TemperatureDeviationCalculatorImpl implements TemperatureDeviationCalculator {
 
   @Override
-  public Deviation calculateDeviation(double temperatureValue, double threshold, double allowedDeviation) {
+  public Deviation calculateDeviation(BigDecimal temperatureValue, double threshold, double allowedDeviation) {
     if (temperatureValue >= (threshold + allowedDeviation)) {
       return Deviation.HIGH;
     }
